+++ cypress/support/setup.ts
@@ -6,18 +6,6 @@ export const setup = (mounter: () => void) => {
   global.process.env = global.process.env || {};
 
   mounter();
-
-  // we cache here so that it is faster
-  cy.fixture('courier-script.txt').then(resp => {
-    // wait for courier script to load
-    cy.intercept(
-      {
-        method: 'GET',
-        url: 'https://courier-components-xvdza5.s3.amazonaws.com/v*.js'
-      },
-      resp
-    );
-  });
   cy.get('#courier-script', { timeout: 25_000 }).should('exist');
 
   // wait for client key to init
