+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "courier-vue-embedded",
-  "version": "4.6.0",
+  "version": "4.5.0",
   "description": "Vue 3/2.7 wrapper for courier",
   "scripts": {
     "dev": "vite",
