+++ src/components/courierPreferencesVue.vue
@@ -9,3 +9,4 @@ import type { PreferencesProps } from '../types/preferences';
 
 const props = defineProps<PreferencesProps>();
 </script>
+../types/preferences.types
