+++ src/index.ts
@@ -2,7 +2,14 @@ import CourierInboxVue from './components/courierInboxVue.vue';
 import CourierToastVue from './components/courierToastVue.vue';
 import CourierPreferencesVue from './components/courierPreferencesVue.vue';
 
-// import { Courier } from './ts/courier';
-import { useCourier } from './ts/useCourier';
+import { useCourier, useCourierInbox, useCourierToast, useCourierPreferences } from './ts/useCourier';
 
-export { CourierInboxVue, CourierToastVue, CourierPreferencesVue, useCourier };
+export {
+  CourierInboxVue,
+  CourierToastVue,
+  CourierPreferencesVue,
+  useCourier,
+  useCourierInbox,
+  useCourierToast,
+  useCourierPreferences
+};
