+++ src/ts/inboxClient.ts
@@ -1,98 +1,20 @@
-import { CourierSDK } from '../types/courier';
-import { InboxProps } from '../types/inbox';
-import { Deferred } from './helpers';
-
-export type InboxView = 'messages' | 'preferences';
-
-export interface IActionElemental {
-  background_color?: string;
-  content: string;
-  data?: Record<string, any>;
-  href: string;
-  type: 'text';
-}
-
-export interface IInboxMessagePreview {
-  actions?: IActionElemental[];
-  archived?: string;
-  created: string;
-  data?: Record<string, any>;
-  messageId: string;
-  opened?: string;
-  preview?: string;
-  read?: string;
-  tags?: string[];
-  title?: string;
-}
-
-export interface IElementalInbox {
-  tenantId?: string;
-  brand?: Brand;
-  from?: number;
-  isLoading?: boolean;
-  isOpen?: boolean;
-  lastMessagesFetched?: number;
-  messages?: Array<IInboxMessagePreview>;
-  startCursor?: string;
-  unreadMessageCount?: number;
-  view?: InboxView;
-}
-
-export type IElementalInboxMessage = IInboxMessage;
-export type IElementalInboxMessagePreview = IInboxMessagePreview;
-
-export interface FetchMessagesDonePayload {
-  messages: IElementalInboxMessagePreview[];
-  appendMessages?: boolean;
-  startCursor?: string;
-}
-
-export interface FetchUnreadMessageCountPayload {
-  count: number;
-}
-
-export interface MarkAllReadDonePayload {
-  ids: string[];
-}
-
-export interface IGetMessagesParams {
-  tenantId?: string;
-  from?: number;
-  isRead?: boolean;
-  limit?: number;
-  tags?: string[];
-}
-
-export interface IFetchMessagesParams {
-  params?: IGetMessagesParams;
-  after?: string;
-}
-
-export interface InboxSdk {
-  fetchMessages: (params?: IFetchMessagesParams) => void;
-  getUnreadMessageCount: (params?: IGetMessagesParams) => void;
-  init: (inbox?: IInbox) => void;
-  markAllAsRead: (fromWS?: boolean) => void;
-  markMessageArchived: (messageId: string, fromWS?: boolean) => Promise<void>;
-  markMessageOpened: (messageId: string, fromWS?: boolean) => Promise<void>;
-  markMessageRead: (messageId: string, fromWS?: boolean) => Promise<void>;
-  markMessageUnread: (messageId: string, fromWS?: boolean) => Promise<void>;
-  newMessage: (transportMessage: IInboxMessagePreview) => void;
-  resetLastFetched: () => void;
-  setView: (view: InboxView) => void;
-  toggleInbox: (isOpen?: boolean) => void;
-  unpinMessage: (messageId: string, fromWS?: boolean) => Promise<void>;
-  trackClick: (messageId: string, trackingId: string) => Promise<void>;
-  setConfig: (config: InboxProps) => void;
-  mergeConfig: (config: InboxProps) => void;
-}
+import type { CourierSDK } from '../types/courier';
+import type {
+  FetchMessagesDonePayload,
+  FetchUnreadMessageCountPayload,
+  IFetchMessagesParams,
+  IGetMessagesParams,
+  IInboxMessagePreview,
+  InboxProps,
+  InboxView
+} from '../types/inbox';
+import { Deferred, withoutUndefinedValues } from './helpers';
 
 export class InboxClient {
   private onceReady = Deferred<void>();
 
   private _fetchMessagesDefered = Deferred<FetchMessagesDonePayload>();
   private _fetchUnreadMessageCountDefered = Deferred<FetchUnreadMessageCountPayload>();
-  private _markAllReadDefered = Deferred<MarkAllReadDonePayload>();
 
   private get sdk(): CourierSDK {
     return window.courier;
@@ -101,7 +23,6 @@ export class InboxClient {
   init() {
     this.sdk.on('inbox/init', () => {
       this.onceReady.resolve();
-      console.dir(window.courier.inbox);
     });
 
     this.sdk.on('inbox/FETCH_MESSAGES/DONE', resp => {
@@ -110,21 +31,6 @@ export class InboxClient {
 
     this.sdk.on('inbox/FETCH_UNREAD_MESSAGE_COUNT/DONE', resp => {
       this._fetchUnreadMessageCountDefered.resolve(resp.payload);
-    });
-
-    // TODO: is this actually called?
-    this.sdk.on('inbox/MARK_ALL_READ/DONE', resp => {
-      this._fetchUnreadMessageCountDefered.resolve(resp.payload);
-    });
-
-    // TODO: does this actually happen?
-    this.sdk.on('inbox/MARK_ALL_READ/ERROR', resp => {
-      console.log('--- error', resp);
-    });
-
-    this.sdk.on('inbox/TOGGLE_INBOX', resp => {
-      // TODO: make this emit events from a vue component
-      console.log('toggle', resp);
     });
   }
 
@@ -143,11 +49,8 @@ export class InboxClient {
   }
 
   async markAllAsRead(fromWS?: boolean) {
-    // TODO: there might be a bug for this one
     await this.onceReady;
-    this._markAllReadDefered = Deferred<MarkAllReadDonePayload>();
     this.sdk.inbox.markAllAsRead(fromWS);
-    return await this._markAllReadDefered;
   }
 
   async markMessageArchived(messageId: string, fromWS?: boolean) {
@@ -165,7 +68,7 @@ export class InboxClient {
     await this.sdk.inbox?.markMessageRead(messageId, fromWS);
   }
 
-  async newMessage(message: IInboxMessagePreview) {
+  async addMessage(message: IInboxMessagePreview) {
     await this.onceReady;
     this.sdk.inbox.newMessage(message);
   }
@@ -180,7 +83,7 @@ export class InboxClient {
     this.sdk.inbox.setView(view);
   }
 
-  async toggleInbox() {
+  async toggleVisibility() {
     await this.onceReady;
     this.sdk.inbox.toggleInbox();
   }
@@ -207,11 +110,13 @@ export class InboxClient {
 
   async setConfig(config: InboxProps) {
     await this.onceReady;
-    this.sdk.inbox.setConfig(config);
+    const toMerge = withoutUndefinedValues(config);
+    this.sdk.inbox.setConfig(toMerge);
   }
 
   async mergeConfig(config: InboxProps) {
     await this.onceReady;
-    this.sdk.inbox.mergeConfig(config);
+    const toMerge = withoutUndefinedValues(config);
+    this.sdk.inbox.mergeConfig(toMerge);
   }
 }
