+++ src/ts/useCourier.ts
@@ -1,22 +1,33 @@
-import { CourierConfig } from '../types/courier';
-import { CourierClient } from './courier';
+import type { CourierConfig } from '../types/courier';
+import { CourierClient } from './courierClient';
 
-let client: CourierClient = null!;
+const client = new CourierClient();
 
-export const useCourier = (config: CourierConfig) => {
-  if (!client) {
-    client = new CourierClient();
-    client.init(config);
+export const useCourier = (initConfig?: CourierConfig) => {
+  if (initConfig) {
+    client.init(initConfig);
   }
 
-  // console.dir(client);
-
-  // refresh token, etch
   return {
-    client
+    init: (config: CourierConfig) => client.init(config),
+    renewSession: (token: string) => client.renewSession(token),
+    on: (action: string, callback: (payload: any) => void) => client.on(action, callback),
+    onAny: (callback: (payload: any) => void) => client.onAny(callback),
+    whenReady: (callback: () => void) => client.whenReady(callback),
+    inbox: useCourierInbox(),
+    toast: useCourierToast(),
+    preferences: useCourierPreferences()
   };
 };
 
-// export const useCourierPreferences = () => {
+export const useCourierInbox = () => {
+  return client.inbox;
+};
 
-// }
+export const useCourierToast = () => {
+  return client.toast;
+};
+
+export const useCourierPreferences = () => {
+  return client.preferences;
+};
